requirements/base.txt
# SHA1:78e8c5e859c020c58c77e59d0552c15952c45077
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
joblib==1.1.0
    # via scikit-learn
numpy==1.22.0
    # via
    #   pandas
    #   scikit-learn
    #   scipy
pandas==1.3.5
    # via -r requirements/base.in
python-dateutil==2.8.2
    # via pandas
pytz==2021.3
    # via pandas
pyyaml==6.0
    # via -r requirements/base.in
scikit-learn==1.0.2
    # via -r requirements/base.in
scipy==1.7.3
    # via scikit-learn
six==1.16.0
    # via python-dateutil
threadpoolctl==3.0.0
    # via scikit-learn
